+++ src/refactoring/features/admin/index.ts
@@ -1,0 +1,4 @@
+export { NewProductForm } from './ui/NewProductForm';
+export { ProductList } from './ui/ProductList';
+export { CouponList } from './ui/CouponList';
+export { NewCouponForm } from './ui/NewCouponForm';
