--- src/refactoring/features/cart/ui/index.ts
@@ -1,4 +1,0 @@
-export { ProductItem } from './ProductItem';
-export { CartItemList } from './CartItemList';
-export { CouponSection } from './CouponSection';
-export { OrderSummary } from './OrderSummary';
